requirements.txt
langchain==0.3.26
langchain-community==0.3.26
flask==3.1.1
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.1.0
faiss-cpu==1.9.0.post1
transformers==4.41.2
accelerate==0.31.0
-e .